# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KeepOn Legacy API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/clientSessions/invite": {
      "post": {
        "operationId": "postClientSessionsInvite",
        "tags": ["clientSessions"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientId": {},
                  "sessionId": {}
                },
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "status", "message", "error", "type", "title"],
        "properties": {
          "code": {
            "type": "number"
          },
          "status": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "required": ["statusCode", "message"],
            "properties": {
              "statusCode": {
                "type": "number"
              },
              "message": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": ["string", "null"]
          }
        },
        "additionalProperties": false
      }
    }
  }
}